+++ src/xviii/Core/GameState_SelectNations.cpp
@@ -41,7 +41,7 @@ uiView{sf::FloatRect({}, {}, xResolution, yResolution)}
 	flagMenuItems.push_back({game->mTextureManager.getSprite(TextureManager::Flag::VEN), Player::Nation::VEN});
 	flagMenuItems.push_back({game->mTextureManager.getSprite(TextureManager::Flag::WAL), Player::Nation::WAL});
 
-	for (int i{0}; i < flagMenuItems.size(); ++i){
+	for (size_t i{0}; i < flagMenuItems.size(); ++i){
 		int spriteXPos = (i * 75);
 		int spriteYPos = yResolution / 2;
 
@@ -82,7 +82,7 @@ void GameState_SelectNations::getInput(){
 					flagIterator = flagMenuItems.begin() + flagMenuItems.size() / 2;
 
 					//And fix the positions of the rest
-					for (int i{0}; i < flagMenuItems.size(); ++i){
+					for (size_t i{0}; i < flagMenuItems.size(); ++i){
 						int spriteXPos = (i * 75);
 						int spriteYPos = yResolution / 2;
